ggml-cpu : fix im2col stride truncation
Forward (f32/f16) and backward (f32) im2col stored tensor strides in
int variables, truncating int64_t nb values. Large strides wrap around
and cause OOB reads/writes into src1/dst data.

Widen ofs0/ofs1 to int64_t in all three variants.

+++ ggml/src/ggml-cpu/ops.cpp
@@ -6381,8 +6381,8 @@ static void ggml_compute_forward_im2col_f32(
     const int64_t OH = is_2D ? ne2 : 1;
     const int64_t OW = ne1;
 
-    int ofs0 = is_2D ? nb13 : nb12;
-    int ofs1 = is_2D ? nb12 : nb11;
+    int64_t ofs0 = is_2D ? nb13 : nb12;
+    int64_t ofs1 = is_2D ? nb12 : nb11;
 
     GGML_ASSERT(nb10 == sizeof(float));
 
@@ -6457,8 +6457,8 @@ static void ggml_compute_forward_im2col_f16(
     const int64_t OH = is_2D ? ne2 : 1;
     const int64_t OW = ne1;
 
-    int ofs0 = is_2D ? nb13 : nb12;
-    int ofs1 = is_2D ? nb12 : nb11;
+    int64_t ofs0 = is_2D ? nb13 : nb12;
+    int64_t ofs1 = is_2D ? nb12 : nb11;
 
     GGML_ASSERT(nb10 == ggml_type_size(src1->type));
 
@@ -6559,8 +6559,8 @@ void ggml_compute_forward_im2col_back_f32(
     const int64_t OH = is_2D ? ne02 : 1;
     const int64_t OW = ne01;
 
-    int ofs0 = is_2D ? nb3 : nb2;
-    int ofs1 = is_2D ? nb2 : nb1;
+    int64_t ofs0 = is_2D ? nb3 : nb2;
+    int64_t ofs1 = is_2D ? nb2 : nb1;
 
     GGML_ASSERT(nb0  == sizeof(float));
